src/ufdr_mounter/ufdr_mounter/requirements.txt

fusepy===3.0.1
